+++ mp3organizer/utils.py
@@ -1,0 +1,23 @@
+import json
+import os
+from pathlib import Path
+
+CACHE_DIR = Path(os.path.expanduser("~")) / ".cache_mp3"
+CACHE_DIR.mkdir(exist_ok=True)
+CACHE_FILE = CACHE_DIR / "cache.json"
+
+
+def load_cache():
+    if CACHE_FILE.exists():
+        with open(CACHE_FILE, "r", encoding="utf-8") as f:
+            return json.load(f)
+    return {}
+
+
+def save_cache(cache):
+    with open(CACHE_FILE, "w", encoding="utf-8") as f:
+        json.dump(cache, f)
+
+
+def sanitize_filename(name: str) -> str:
+    return ''.join(c for c in name if c.isalnum() or c in (' ', '-', '_')).title()
